requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==46.0.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fabric==3.2.2
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.12.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
hf-cli==0.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.36.1
importlib_metadata==7.2.1
importlib_resources==6.5.2
invoke==2.2.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==2.15.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml-dtypes==0.2.0
namex==0.0.8
narwhals==1.41.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.14
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==4.24.4
psutil==6.1.1
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyNaCl==1.6.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
random2==1.0.2
referencing==0.35.1
regex==2024.11.6
reportlab==4.4.4
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
ruff==0.9.1
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.4.0
scipy==1.13.1
scp==0.15.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.41.3
streamlit==1.31.1
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
tf-keras==2.15.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tinycss2==1.4.0
tokenizers==0.22.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.5.0
tzdata==2024.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
whitenoise==6.5.0
widgetsnbextension==4.0.14
wrapt==1.14.1
zipp==3.21.0
